requirements.txt
lxml==4.3.4
pdf2image==1.12.0
Pillow==6.1.0
python-pptx==0.6.18
XlsxWriter==1.1.8
